Restructure with DDD layers, center generator, remove logos
