--- src/App.jsx
@@ -1,24 +1,0 @@
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
-import './App.css'
-import TaleGenerator from './components/TaleGenerator.jsx'
-
-function App() {
-
-  return (
-    <>
-      <div>
-        <a href="https://vite.dev" target="_blank">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Image to Tale</h1>
-      <TaleGenerator />
-    </>
-  )
-}
-
-export default App
