+++ src/application/generateTale.js
@@ -1,0 +1,4 @@
+export default async function generateTale(storyGenerator, image, context) {
+  if (!storyGenerator) throw new Error('StoryGenerator is required')
+  return storyGenerator.generate(image, context)
+}
