+++ src/domain/StoryGenerator.js
@@ -1,0 +1,6 @@
+export default class StoryGenerator {
+  // eslint-disable-next-line no-unused-vars
+  async generate(_image, _context) {
+    throw new Error('Not implemented')
+  }
+}
