+++ src/presentation/App.jsx
@@ -1,0 +1,14 @@
+import './App.css'
+import TaleGenerator from './components/TaleGenerator.jsx'
+
+function App() {
+
+  return (
+    <>
+      <h1>Image to Tale</h1>
+      <TaleGenerator />
+    </>
+  )
+}
+
+export default App
