+++ src/presentation/components/TaleGenerator.module.css
